+++ packages/generator-widget/CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
--   We upgraded React Native to version 0.77.3 for generated widgets.
+-   We upgraded React Native to version 0.78.2 for generated widgets.
 
 -   We updated the required Node.js version to 20 or newer for the generator and generated widget projects.